+++ packages/sdks/react-sdk/CHANGELOG.md
@@ -2,6 +2,47 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.28.7](https://github.com/descope/descope-js/compare/react-sdk-2.28.6...react-sdk-2.28.7) (2026-06-01)
+
+### Dependency Updates
+
+* `role-management-widget` updated to version `0.7.0`
+## [2.28.6](https://github.com/descope/descope-js/compare/react-sdk-2.28.5...react-sdk-2.28.6) (2026-05-28)
+
+### Dependency Updates
+
+* `user-management-widget` updated to version `0.13.1`
+* `user-profile-widget` updated to version `0.11.3`
+* `outbound-applications-widget` updated to version `0.3.3`
+* `web-component` updated to version `3.65.0`
+* `tenant-profile-widget` updated to version `0.6.3`
+## [2.28.5](https://github.com/descope/descope-js/compare/react-sdk-2.28.4...react-sdk-2.28.5) (2026-05-28)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.6.2`
+* `audit-management-widget` updated to version `0.7.2`
+* `role-management-widget` updated to version `0.6.2`
+* `user-management-widget` updated to version `0.13.0`
+* `user-profile-widget` updated to version `0.11.2`
+* `applications-portal-widget` updated to version `0.7.2`
+* `outbound-applications-widget` updated to version `0.3.2`
+* `web-component` updated to version `3.64.3`
+* `tenant-profile-widget` updated to version `0.6.2`
+## [2.28.4](https://github.com/descope/descope-js/compare/react-sdk-2.28.3...react-sdk-2.28.4) (2026-05-26)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.6.1`
+* `audit-management-widget` updated to version `0.7.1`
+* `role-management-widget` updated to version `0.6.1`
+* `user-management-widget` updated to version `0.12.1`
+* `user-profile-widget` updated to version `0.11.1`
+* `applications-portal-widget` updated to version `0.7.1`
+* `outbound-applications-widget` updated to version `0.3.1`
+* `web-component` updated to version `3.64.2`
+* `web-js-sdk` updated to version `1.50.1`
+* `tenant-profile-widget` updated to version `0.6.1`
 ## [2.28.3](https://github.com/descope/descope-js/compare/react-sdk-2.28.2...react-sdk-2.28.3) (2026-05-25)
 
 ### Dependency Updates